requirements.txt
Flask==2.2.5
Pillow==9.4.0
rembg==2.0.25
Werkzeug==2.2.3